day03 part 1
What a terrible experience.

I have spent hours debugging this, because I was convinced that the
website had told me my answer was too high. There are literally a dozen
test cases from the internet, all of which passed. Manual inspection of
special-purpose debug code, which succeeded.

And then, it must have been a copy-paste error of some kind???
Because the answer I've been getting all along is right, now.