day12: input parsing and minor calculation
So, the max unknown bits in any condition record is 18, which means
there are 262144 possibilities to check for that record if we
brute-force it.

I could try to do something clever with dynamic programming, but for
now, I'm leaning toward just checking every possibility by creating
some kind of bitwise mapping over the unknown bits.